Fix issue with widget's contents being too high on clicking in preview.
When clicking a widget in the preview, the 'onChangeExpanded' method is called.
This included a jQuery method that adds a 'margin-top' value.
Remove this method call, as it caused a spacing issue.
The inactive panels no longer occupy space, from commit 8894171c.
So there's no adjustment needed for the controls to appear properly.

Fixes #34747.