Customize : customize-base.js QUnit test setup and teardown.
Define same 'setupAndTearDown' function as in commit 63791f8 for customize-controls.js.
This can't be abstracted into a common file, as it alters the module-scope variable api.
This function avoids altering the global 'window.wp' object by cloning it to 'api'.

Also, fix alignment of declared variables in 'customize-control.js.'

Fixes #35857.