Customize: Test 'Selective Refresh' feature by altering settings.
Mock settings for the 'blogname' and 'description,' and change them.
Test that they are refreshed properly.
Add more tests for methods and properties of api.selectiveRefresh.
Modify the .html file with the markup for the settings.
Remove the 'sinon.js' and 'sinon-qunit.js' files.
They seemed to interfere with the 'setTimeout' function.
Also, undbind the 'customize-preview.js' tests from an event after they finish.

Fixes #35857.