Customize : Menu and widget search field UI changes.
Create search icon in <input>, using neighbor <div> with :after.
It's not possible to set this as an :after element on the <input>.
Increase padding-left from 10px to 30px to create space for this.
Mainly copy styling of .clear-results in customize-nav-menus.css#L626.
In menu search field, increase clickable area of x icon.

Fixes #36908.