Menus: edit.php - fix issue with ajax requests not completing.
As Weston pointed out, ajax requests sometimes didn't complete.
This is because they're debounced.
And the submit form button reloads the page before they finish.
So pass true as the 'immediate' argument to _.debounce.
This submits the first ajax request, without a 2000 ms delay.
Still  this won't help with clicking many of the screen option boxes.

Fixes #36202.