Customize: Display checkmark on newly-created pages.
When adding a new page to page to a menu,
The page uses the template available-menu-item.
So pass this template a value of true for 'is_new_item'.
The template will then output 'item-added'.
Existing pages also use this template.
So this can't simply ouput the class for all of them.

Fixes #38499.